Renamed pde_solver module to time_stepper.
